requirements.txt
coloredlogs
dacite
dataclasses_json
Flask==1.1.1
Flask-Cors==3.0.8
google-cloud-bigquery
google-cloud-firestore
grpcio
mock-firestore
pandas
pylint
pytest
pytest-flask
requests
xmltodict
zeep